requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
langchain>=0.1.0
openai>=1.0.0

# Vector Database
chromadb>=0.4.0
faiss-cpu>=1.7.4

# OCR and Document Processing
pytesseract>=0.3.10
pdf2image>=1.16.3
PyPDF2>=3.0.1
opencv-python>=4.8.0
pillow>=10.0.0
python-docx>=0.8.11

# Web Framework and UI
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
tqdm>=4.65.0
requests>=2.31.0

# Testing & Dev
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
